Split semantic language service responsibilities
